added custom feed and net share support (#7737)
* added custom feed and net share support

* fixed invocation

* added ui option

* made some fixes

* fixed list not working by package id on custom fields

* finished adding tests

* resolved telemetry related comments

* fixed telemetry addition

* fixed event name

+++ Tasks/VsTestPlatformToolInstallerV1/helpers.ts
@@ -11,10 +11,14 @@ export function setVsTestToolLocation(toolPath: string) {
 }
 
 export function cleanUpTempConfigFile(tempConfigFilePath: string) {
+    if (isNullEmptyOrUndefined(tempConfigFilePath)) {
+        return;
+    }
+
     try {
         fs.unlinkSync(tempConfigFilePath);
     } catch (error) {
-        tl.debug(`Failed to delete temp confi file ${tempConfigFilePath} with error ${error}.`);
+        tl.debug(`Failed to delete temp config file ${tempConfigFilePath} with error ${error}.`);
     }
 }
